Add jest mock, split out files, update entry point
